+++ modules/integration/tests-integration/tests-backend/src/test/java/org/wso2/identity/integration/test/oauth2/OAuth2ServiceErrorResponseTest.java
@@ -18,8 +18,13 @@ package org.wso2.identity.integration.test.oauth2;
 import org.apache.http.Header;
 import org.apache.http.HttpResponse;
 import org.apache.http.NameValuePair;
+import org.apache.http.client.entity.UrlEncodedFormEntity;
+import org.apache.http.client.methods.HttpPost;
 import org.apache.http.impl.client.DefaultHttpClient;
 import org.apache.http.message.BasicNameValuePair;
+import org.apache.http.util.EntityUtils;
+import org.json.simple.JSONObject;
+import org.json.simple.JSONValue;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
@@ -29,19 +34,21 @@ import org.wso2.carbon.identity.oauth.stub.dto.OAuthConsumerAppDTO;
 import org.wso2.identity.integration.test.utils.OAuth2Constant;
 import org.wso2.carbon.integration.common.admin.client.AuthenticatorClient;
 
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
 
 public class OAuth2ServiceErrorResponseTest extends OAuth2ServiceAbstractIntegrationTest {
 
-    private AuthenticatorClient logManger;
+    private AuthenticatorClient authenticatorClient;
     private DefaultHttpClient client;
 
     @BeforeClass(alwaysRun = true)
     public void testInit() throws Exception {
         super.init(TestUserMode.SUPER_TENANT_USER);
-        logManger = new AuthenticatorClient(backendURL);
-        logManger.login(isServer.getSuperTenant().getTenantAdmin().getUserName(),
+        authenticatorClient = new AuthenticatorClient(backendURL);
+        authenticatorClient.login(isServer.getSuperTenant().getTenantAdmin().getUserName(),
                                               isServer.getSuperTenant().getTenantAdmin().getPassword(),
                                               isServer.getInstance().getHosts().get("default"));
 
@@ -52,7 +59,7 @@ public class OAuth2ServiceErrorResponseTest extends OAuth2ServiceAbstractIntegra
     @AfterClass(alwaysRun = true)
     public void atEnd() throws Exception {
         removeOAuthApplicationData();
-        logManger = null;
+        authenticatorClient = null;
         consumerKey = null;
     }
 
@@ -66,9 +73,9 @@ public class OAuth2ServiceErrorResponseTest extends OAuth2ServiceAbstractIntegra
         consumerSecret = appDto.getOauthConsumerSecret();
     }
 
-    @Test(groups = "wso2.is", description = "Test unsupported grant type error response", dependsOnMethods =
-            "testRegisterApplication")
-    public void testUnsupportedGrantTypeErrorResponse() throws Exception {
+    @Test(groups = "wso2.is", description = "check whether the client is not authorized to use the " +
+            "authorization grant type", dependsOnMethods = "testRegisterApplication")
+    public void testUnauthorizeClientErrorResponse() throws Exception {
         List<NameValuePair> urlParameters = new ArrayList<NameValuePair>();
         urlParameters.add(new BasicNameValuePair("response_type",
                                                  OAuth2Constant.OAUTH2_GRANT_TYPE_IMPLICIT));
@@ -89,6 +96,36 @@ public class OAuth2ServiceErrorResponseTest extends OAuth2ServiceAbstractIntegra
         Assert.assertTrue(locationURI.contains("The+authenticated+client+is+not+authorized+to+use+this+authorization+" +
                 "grant+type"));
     }
+
+    @Test(groups = "wso2.is",
+            description = "Check whether the grant type is not supported by the authorization server",
+            dependsOnMethods = "testRegisterApplication")
+    public void testUnsupportedGrantype() throws Exception {
+
+        String unSupportedGrantType = "1qaz2wsx";
+        client = new DefaultHttpClient();
+        ArrayList<NameValuePair> postParameters;
+        HttpPost httpPost = new HttpPost(OAuth2Constant.ACCESS_TOKEN_ENDPOINT);
+        //generate post request
+        httpPost.setHeader("Authorization", "Basic " + getBase64EncodedString(consumerKey, consumerSecret));
+        httpPost.setHeader("Content-Type", "application/x-www-form-urlencoded");
+        postParameters = new ArrayList<NameValuePair>();
+        postParameters.add(new BasicNameValuePair("username", "admin"));
+        postParameters.add(new BasicNameValuePair("password", "admin"));
+        postParameters.add(new BasicNameValuePair("grant_type", unSupportedGrantType));
+        httpPost.setEntity(new UrlEncodedFormEntity(postParameters));
+        HttpResponse response = client.execute(httpPost);
+
+        BufferedReader rd = new BufferedReader(new InputStreamReader(response.getEntity().getContent()));
+
+        Object obj = JSONValue.parse(rd);
+        String errorMessage = ((JSONObject) obj).get("error").toString();
+        EntityUtils.consume(response.getEntity());
+        Assert.assertEquals(OAuth2Constant.UNSUPPORTED_GRANT_TYPE, errorMessage,
+                "Unsupported grant type should have " + "produced error code : "
+                        + OAuth2Constant.UNSUPPORTED_GRANT_TYPE);
+    }
+
 
     public OAuthConsumerAppDTO createApplication() throws Exception {
